backend/requirements.txt
# 网文生成 AI Agent 后端 - 三大核心数据库交互
# Python 3.10+

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# SQLAlchemy 异步 + SQLite
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0

# Neo4j 图数据库（官方驱动，支持异步）
neo4j>=5.16.0

# 可选：环境变量
python-dotenv>=1.0.0
